# VMS-2879 review fix
# VMS-5400 review fix
# cctu listen improvement
# minor bugfix in RetryTimer
#CLOUD-312 Fixed some bugs in cloud_db load testing utility
#CLOUD-838 Fixed bug in test
#CLOUD-838 UDT: More numeric constants -> enum replacements
#CLOUD-838 UDT: Numeric constants replaced with enum
#CLOUD-916 Added some debug code
#CLOUD-916 Disabled all socket timeouts in blinking test SocketUdt.TransferAsync
#CLOUD-916 Fixed UdtSocket::handle implementation
#CLOUD-916 Fixed issue that caused UDT connect to fail sometimes.
#CLOUD-916 Replacing numeric values with enum
#CLOUD-922 Added account registration/activation timestamps
#CLOUD-922 Fixed AccountNewTest.account_timestamps test failure because of round error.
#CLOUD-922 Registering cloud account now requires AccountRegistrationData structure, not AccountData which contains several cloud_db generated values
#CLOUD-922 fixed link error
#CLOUD-923 Method naming fixes.
#CLOUD-923 cloud_db: Added system registration timestamp
#VMS-3877 Replaced std::random_device with not throwing nx::utils::random. Decreased test loop.
#VMS-5437 Call to acceptAsync on socket in a blocking mode results in SystemError::notSupported error.
#VMS-5461 Mediaserver now rejects transaction connections from itself
#VMS-5538 Added more logs
#VMS-5570 Added assert when invoking null function
#add more log warnings on configure rest handler error
#compile error fix
#merge system fix error checking: JSON error didn't checked afer /configure call
Added some debug functions.
Added some important log messages.
Added test validating invited cloud user authenticating in mediaserver
Axis 6045 MaxFPS fixed
Backed out changeset: 7f51d7c6660c. missing new file
Backed out changeset: 85720bd61866
Branch renamed to vms_2.6
CLOUD-562: Rename /ec2/* methods to /cdb/ec2/*
Change method for embedding flash plugin into the page
Clean code and change logic - start playing only when manifest is ready
Close dev_2.6 branch
Close prod_2.6 branch
Close x_vms_4647_tooltips_refactor branch
Cloud account activation code expiration period has been moved to settings
Cloud promo notification displayed to owners only
Coding style fixes
Coding style fixes.
Compile fix
Decreased some messages logging level
Embed chromeless player for testing purposes
Enabled -Werror on connection_mediator_ut
Fixed compile error
Fixed compile warning
Fixed crash on test failure
Fixed numerous compile warnings
Fixed some code formatting in UDT
Fixed test crash on failure
Fixed typo
Fixed undefined behaviour in UdpHolePunching.MultiConnect test
Fixed: Connecting/disconnecting system from cloud is performed based on cloudSystemId/cloudAuthKey values only
Force player to use full URL to request hls
Links panel fix
MOBILE-476: Player keeps playing video for a short period of time after pause was requested
MOBILE-481: Fix some issues in video aspect ratio handling
MOBILE-483: Improve initial timeline scale
MOBILE-483: Increase default timeline window size to 24 hours
NX1 ssl comile fix
NX_VMS-3522: Wix: Translation fixes: removed &amp and formatted hyperlinks
QnConnectToCloudWatcher now can use overrided cloud url (needed for tests)
Removed assert when trying to perform acceptAsync on socket in blocking mode
Removed unused code fragment
Removed unused function
Simplified recent RetryTimer modification
Sorting out mess with gtest custom type printer functions.
Spelling fix
Strings correction
TEST-180 Functional tests bugfix: script must not be expected in current dir
TEST-180 Functional tests: add --vm-name-prefix parameter
TEST-180 Functional tests: add ability to configure boxes; add nat configuration to nat connection test
TEST-180 Functional tests: add debug printing to nat box provision script
TEST-180: merge test
Translations strings fix: system -> System
Translations: Cloud Beta made untranslatable
Typo fixed
Update flashls plugin to latest version
Updated Russian translations
Using nx::utils::utcTime() insead of std::chrono::system_clock::now()
VMS-2052: Fix resource credentials reading from resourceData.json
VMS-2879 add option to specify supported tls protocol version (e.g., allow tls 1.2.0 only)
VMS-3295: Localizations and customizations for webadmin
VMS-4767 - Make previews on tooltips in notifications the same as in tree
VMS-4852: Add vertical lines to the timeline scrollbar
VMS-5219 - Change text in 'motion on camera' email
VMS-5337: Add Rest API breaking changes info to web-admin
VMS-5356: Force motion detection is unusable in case of AV cameras
VMS-5358: 'You have the latest version...' message should not appear until update is finished
VMS-5371: Fill numerus forms in en_US/en_GB translations. Take 2
VMS-5379 - 'Clear Screen' option for VideoWall doesn't remove icon at User Settings until relogin
VMS-5379 - 'Clear Screen' option for VideoWall doesn't remove icon at User Settings untill relogin
VMS-5385: Fix updates widget lock after cancelling update
VMS-5399: Make webadmin use *.json file for language
VMS-5464: Wrong line height in table in Storage Analytics
VMS-5491:Get rid of contractions in source texts
VMS-5501: Add manual camera take too much time
VMS-5508: Add user's name (or login if name is empty) into tooltip in tree
VMS-5510 - Auto-login to cloud systems
VMS-5511 Remove all crash dump files before install update on NX1
VMS-5513: Alt+Space should invoke the main menu
VMS-5515 - Preloader doesn't show during connecting to System
VMS-5518 - NX client was crashed (build 14105)
VMS-5524 Change auto-generated certificate algorithm to SHA2 or SHA256
VMS-5525: Bookmark Log. Filter by date doesn't work on first attempt
VMS-5528: There is no any information when PTZ tour is active
VMS-5532 - Fix 'The Word_from_Capital' translation
VMS-5552: Sometimes motion recording action doesnt work if postRecording is not 0
VMS-5553: New system is merged to cloud system incorrectly
VMS-5561 Sometimes HardwareID can't be calculated
WEB-472: Webadmin localization: refine sources
added Turkish translation
close branch
fixed compile warnings
fixed texts
more spelling
spelling
spelling and acronyms
spelling fix
system_clock::time_point gtest printer now prints nanoseconds, not milliseconds
the system -> System
typo fixed

<==== Mercurial data ====>
Branch: vms_3.1_gui

+++ unit_tests/nx_network_ut/src/http/http_stream_reader_ut.cpp
@@ -373,7 +373,7 @@ TEST( HttpStreamReader, MultipleMessages )
             ++messageNumber;
         }
 
-        ASSERT_EQ( messageNumber, messagesToParse.size() );
+        ASSERT_EQ( (size_t)messageNumber, messagesToParse.size() );
     }
 }
 
@@ -709,6 +709,6 @@ TEST( HttpStreamReader, MultipleRequests )
             ++messageNumber;
         }
 
-        ASSERT_EQ( messageNumber, messagesToParse.size() );
+        ASSERT_EQ( (size_t)messageNumber, messagesToParse.size() );
     }
 }
